# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZOVARK SOC Automation API
  version: 1.2.0
  description: |
    AI-powered SOC automation platform. Manages security investigations,
    task orchestration, SIEM integration, shadow mode automation, kill switch
    controls, token quota management, MCP approval gates, analytics,
    and multi-tenant operations.

    ## Authentication
    - **Access tokens** expire in 15 minutes and are returned in the JSON response body.
    - **Refresh tokens** are stored in httpOnly cookies (7-day lifetime) and are NOT sent as Bearer tokens.
      Use `POST /api/v1/auth/refresh` to obtain a new access token.
    - **API keys** (zovark_ prefix) can be used for M2M authentication via X-API-Key header.
  contact:
    name: ZOVARK Platform
  license:
    name: Proprietary
servers:
  - url: http://localhost:8090
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /api/v1/auth/login, /api/v1/auth/callback,
        or /api/v1/auth/refresh. Access tokens expire in 15 minutes. Refresh tokens
        are delivered via httpOnly cookies and cannot be sent as Bearer tokens — use
        the /api/v1/auth/refresh endpoint to rotate access tokens silently.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for M2M authentication (zovark_ prefix)
...
paths:
...
  /api/v1/ingest/splunk:
    post:
      summary: Ingest alerts from Splunk
      tags: [SIEM Ingest]
      description: |
        Receives alerts from Splunk via webhook. Accepts Splunk alert action
        payload format and converts to ZOVARK internal alert schema. Each alert
        is deduplicated and queued for investigation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                sid:
                  type: string
                  description: Splunk search ID
                search_name:
                  type: string
                  description: Name of the saved search / correlation rule
                result:
                  type: object
                  description: Splunk result row containing alert fields
                  properties:
                    _raw:
                      type: string
                    _time:
                      type: string
                    source:
                      type: string
                    sourcetype:
                      type: string
                    host:
                      type: string
                app:
                  type: string
                  description: Splunk app context
      responses:
        '202':
          description: Alert accepted for investigation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  task_id:
                    type: string
                    format: uuid
                  message:
                    type: string
        '400':
          description: Invalid payload
        '401':
          description: Missing or invalid API key
        '429':
          description: Rate limit exceeded
  /api/v1/ingest/elastic:
    post:
      summary: Ingest alerts from Elastic Security
      tags: [SIEM Ingest]
      description: |
        Receives alerts from Elastic Security via webhook action. Accepts
        Elastic alert payload format (Kibana alerting or Elastic Security
        detection rules) and converts to ZOVARK internal alert schema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alert]
              properties:
                alert:
                  type: object
                  description: Elastic Security alert object
                  properties:
                    id:
                      type: string
                    rule:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        severity:
                          type: string
                    signal:
                      type: object
                      properties:
                        original_event:
                          type: object
                        ancestors:
                          type: array
                          items:
                            type: object
                context:
                  type: object
                  description: Additional context from Elastic
                  properties:
                    space_id:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
      responses:
        '202':
          description: Alert accepted for investigation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  task_id:
                    type: string
                    format: uuid
                  message:
                    type: string
        '400':
          description: Invalid payload
        '401':
          description: Missing or invalid API key
        '429':
          description: Rate limit exceeded
...